+++ src/components/inputs/InputWrapper.vue
@@ -1,7 +1,8 @@
 <template>
   <component
-    :is="inputList[attribute.type]"
+    :is="inputList[attribute.definition.type]"
     :attribute="attribute"
+    :plugin="plugin"
     @update:model-value="(event) => emit('update:attribute', event)"
   />
 </template>
@@ -14,6 +15,10 @@ defineProps({
     type: Object,
     required: true,
   },
+  plugin: {
+    type: Object,
+    required: true,
+  },
 });
 
 const emit = defineEmits(['update:attribute']);
@@ -22,5 +27,6 @@ const inputList = {
   Boolean: defineAsyncComponent(() => import('./BooleanInput')),
   String: defineAsyncComponent(() => import('./StringInput')),
   Number: defineAsyncComponent(() => import('./NumberInput')),
+  Reference: defineAsyncComponent(() => import('./ReferenceInput')),
 };
 </script>